+++ CHANGELOG.md
@@ -21,6 +21,7 @@ https://keepachangelog.com/ja/1.0.0/ に基づいて記述していく
 Unreleased
 -----------
 ### Added
+- add nrm1 and get_residual_l2(Dense) https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/191
 - add Frank matrix creation and eigenvalue calculation routine https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/189
 - add jacobi solver https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/188
 - add jacobi preconditioner https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/188
@@ -34,6 +35,7 @@ Unreleased
 - add makefile target `make oss-cpu` `make oss-cpu` `make mkl-cpu` `make mkl-gpu` https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/178
 
 ### Changed 
+- move equation::solver and equation::precondition to solver::solver and solver::precondition https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/190
 - exclude src/internal doxygen https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/182
 - update allgebra 20.12.2 https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/182
 - include algorithm in internal.hpp https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/181